fix: move AuthenticationRequired to types
